Refactor: change old render function with the new
